Move all crash related tasks into a util class
